release()とprint_usage()以外をsubset.cとして分割（私達にはダイエットが必要です）
